feat: Update for pyjwt >= 2.0.0
Update the version of pyjwt and requests used by this library to be compatible
with other sanely modern versions of python packages. The fact that this has not
been done already is rather absurd.

https://github.com/snowflakedb/snowflake-ingest-python/pull/45 made the severely
myopic choice of pinning pyjwt to 1.7.1 rather than updating the code to be
compatible with pyjwt >= 2.0.0. This directly harms all potential consumers of
the package, as they cannot update their code to be hardened to CVEs or consume
other updates of libraries that have incompatible dependencies with
snowflake-ingest's. In a node/npm style language this would be fine, but
Python's dependency resolution mechanics do not allow for installing or relying
on multiple versions of a particular library.

The only significant breaking change introduced by pyjwt 2.0.0 is that the
encode method no longer returns bytes type strings. By checking the return
value, we can know whether to do an extra decode.